Move grab logic from game.js to grab-manager.js
