+++ io-manager.js
@@ -24,6 +24,7 @@ import transformControls from './transform-controls.js';
 import storyManager from './story.js';
 // import domRenderer from './dom-renderer.jsx';
 import raycastManager from './raycast-manager.js';
+import grabManager from './grab-manager.js';
 
 // const localVector = new THREE.Vector3();
 // const localVector2 = new THREE.Vector3();
@@ -419,7 +420,7 @@ ioManager.keydown = e => {
     case 70: { // F
       e.preventDefault();
       e.stopPropagation();
-      if (game.canPush()) {
+      if (grabManager.canPush()) {
         ioManager.keys.forward = true;
       } else {
         /* if (game.canJumpOff()) {
@@ -446,7 +447,7 @@ ioManager.keydown = e => {
         } else if (game.canBuild()) {
           game.setBuildMode('floor');
         } else { */
-          game.menuDelete();
+        grabManager.menuDelete();
         // }
       }
       break;
@@ -457,7 +458,7 @@ ioManager.keydown = e => {
           game.startBuild('stair');
         } else if (game.canBuild()) {
           game.setBuildMode('stair');
-        } else */if (game.canPush()) {
+        } else */if (grabManager.canPush()) {
           ioManager.keys.backward = true;
         } else {
           ioManager.keys.ctrl = true;
@@ -507,8 +508,8 @@ ioManager.keydown = e => {
       } else {
         game.menuMiddleUp();
 
-        if (game.canRotate()) {
-          game.menuRotate(-1);
+        if (grabManager.canRotate()) {
+          grabManager.menuRotate(-1);
         } else {
           game.menuActivateDown();
         }
@@ -537,8 +538,8 @@ ioManager.keydown = e => {
     }
     case 82: { // R
       if (cameraManager.pointerLockElement) {
-        if (game.canRotate()) {
-          game.menuRotate(1);
+        if (grabManager.canRotate()) {
+          grabManager.menuRotate(1);
         } else {
           game.dropSelectedApp();
         }
@@ -608,7 +609,7 @@ ioManager.keydown = e => {
       break;
     }
     case 192: { // tilde
-      game.toggleEditMode();
+      grabManager.toggleEditMode();
       break;
     }
   }
@@ -755,7 +756,7 @@ ioManager.mouseleave = e => {
 };
 ioManager.click = e => {
   if (cameraManager.pointerLockElement) {
-    game.menuClick(e);
+    grabManager.menuClick(e);
   } else {
     // game.setContextMenu(false);
